Work around '#' escaping bug in bip21 crate
bip21 dependency temporarily specified to use bugfix PR branch

The `pj` parameter of the BIP 21 URL is itself a URL which contains a
fragment.

This is not escaped by bip21 during serialization, which according to
RFC 3986 truncates the `pj` parameter, making everything that follows
part of the fragment.

Deserialization likewise ignores #, parsing it as part of the value
which round trips with the incorrect serialization behavior.

Temporarily depend on bugfix pr branch for bip21